refactor for internal types and `IndexedDbQueryBuilder`
